moved thumbnail widget to separated file
